+++ docs/python/configure-asgi-middleware.py
@@ -1,6 +1,6 @@
 # Broken load order, only used for linting
 from channels.routing import ProtocolTypeRouter, URLRouter
-from reactpy_django import REACTPY_WEBSOCKET_PATH
+from reactpy_django import REACTPY_WEBSOCKET_ROUTE
 
 django_asgi_app = ""
 
@@ -15,7 +15,7 @@ application = ProtocolTypeRouter(
         "websocket": SessionMiddlewareStack(
             AuthMiddlewareStack(
                 URLRouter(
-                    [REACTPY_WEBSOCKET_PATH],
+                    [REACTPY_WEBSOCKET_ROUTE],
                 )
             )
         ),