+++ src/js/src/index.js
@@ -1,34 +1,59 @@
 import { mountLayoutWithWebSocket } from "@reactpy/client";
 
 // Set up a websocket at the base endpoint
-const LOCATION = window.location;
+let HTTP_PROTOCOL = window.location.protocol;
 let WS_PROTOCOL = "";
-if (LOCATION.protocol == "https:") {
-	WS_PROTOCOL = "wss://";
+if (HTTP_PROTOCOL == "https:") {
+	WS_PROTOCOL = "wss:";
 } else {
-	WS_PROTOCOL = "ws://";
+	WS_PROTOCOL = "ws:";
 }
-const WS_ENDPOINT_URL = WS_PROTOCOL + LOCATION.host + "/";
 
 export function mountViewToElement(
 	mountElement,
-	reactpyWebsocketUrl,
-	reactpyWebModulesUrl,
-	maxReconnectTimeout,
-	componentPath
+	reactpyHost,
+	reactpyUrlPrefix,
+	reactpyReconnectMax,
+	reactpyComponentPath,
+	reactpyResolvedWebModulesPath
 ) {
-	const WS_URL = WS_ENDPOINT_URL + reactpyWebsocketUrl + componentPath;
-	const WEB_MODULE_URL = LOCATION.origin + "/" + reactpyWebModulesUrl;
+	// Determine the Websocket route
+	let wsOrigin;
+	if (reactpyHost) {
+		wsOrigin = `${WS_PROTOCOL}//${reactpyHost}`;
+	} else {
+		wsOrigin = `${WS_PROTOCOL}//${window.location.host}`;
+	}
+	const websocketUrl = `${wsOrigin}/${reactpyUrlPrefix}/${reactpyComponentPath}`;
+
+	// Determine the HTTP route
+	let httpOrigin;
+	let webModulesPath;
+	if (reactpyHost) {
+		httpOrigin = `${HTTP_PROTOCOL}//${reactpyHost}`;
+		webModulesPath = `${reactpyUrlPrefix}/web_module`;
+	} else {
+		httpOrigin = `${HTTP_PROTOCOL}//${window.location.host}`;
+		if (reactpyResolvedWebModulesPath) {
+			webModulesPath = reactpyResolvedWebModulesPath;
+		} else {
+			webModulesPath = `${reactpyUrlPrefix}/web_module`;
+		}
+	}
+	const webModuleUrl = `${httpOrigin}/${webModulesPath}`;
+
+	// Function that loads the JavaScript web module, if needed
 	const loadImportSource = (source, sourceType) => {
 		return import(
-			sourceType == "NAME" ? `${WEB_MODULE_URL}${source}` : source
+			sourceType == "NAME" ? `${webModuleUrl}/${source}` : source
 		);
 	};
 
+	// Start rendering the component
 	mountLayoutWithWebSocket(
 		mountElement,
-		WS_URL,
+		websocketUrl,
 		loadImportSource,
-		maxReconnectTimeout
+		reactpyReconnectMax
 	);
 }