+++ src/reactpy_django/exceptions.py
@@ -4,3 +4,7 @@ class ComponentParamError(TypeError):
 
 class ComponentDoesNotExistError(AttributeError):
     ...
+
+
+class InvalidHostError(ValueError):
+    ...