+++ tests/test_app/asgi.py
@@ -11,7 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "test_app.settings")
 
 # Fetch ASGI application before importing dependencies that require ORM models.
@@ -20,15 +19,13 @@ http_asgi_app = get_asgi_application()
 from channels.auth import AuthMiddlewareStack  # noqa: E402
 from channels.routing import ProtocolTypeRouter, URLRouter  # noqa: E402
 from channels.sessions import SessionMiddlewareStack  # noqa: E402
-
-from reactpy_django import REACTPY_WEBSOCKET_PATH  # noqa: E402
-
+from reactpy_django import REACTPY_WEBSOCKET_ROUTE  # noqa: E402
 
 application = ProtocolTypeRouter(
     {
         "http": http_asgi_app,
         "websocket": SessionMiddlewareStack(
-            AuthMiddlewareStack(URLRouter([REACTPY_WEBSOCKET_PATH]))
+            AuthMiddlewareStack(URLRouter([REACTPY_WEBSOCKET_ROUTE]))
         ),
     }
 )