+++ tests/test_app/settings.py
@@ -40,7 +40,6 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "channels",  # Websocket library
     "reactpy_django",  # Django compatiblity layer for ReactPy
     "test_app",  # This test application
 ]
@@ -161,6 +160,10 @@ STATICFILES_FINDERS = [
 ]
 
 # Logging
+LOG_LEVEL = "WARNING"
+if DEBUG and ("test" not in sys.argv):
+    LOG_LEVEL = "DEBUG"
+
 LOGGING = {
     "version": 1,
     "disable_existing_loggers": False,
@@ -172,7 +175,7 @@ LOGGING = {
     "loggers": {
         "reactpy_django": {
             "handlers": ["console"],
-            "level": "DEBUG" if DEBUG else "WARNING",
+            "level": LOG_LEVEL,
         },
     },
 }
@@ -180,4 +183,4 @@ LOGGING = {
 
 # ReactPy Django Settings
 REACTPY_AUTH_BACKEND = "django.contrib.auth.backends.ModelBackend"
-REACTPY_BACKHAUL_THREAD = "test" not in sys.argv
+REACTPY_BACKHAUL_THREAD = "test" not in sys.argv and "runserver" not in sys.argv