+++ tests/test_app/views.py
@@ -1,6 +1,8 @@
 import inspect
+from itertools import cycle
 
 from channels.db import database_sync_to_async
+from django.http import HttpRequest
 from django.shortcuts import render
 from django.views.generic import TemplateView, View
 
@@ -9,6 +11,35 @@ from .types import TestObject
 
 def base_template(request):
     return render(request, "base.html", {"my_object": TestObject(1)})
+
+
+def host_port_template(request: HttpRequest, port: int):
+    host = request.get_host().replace(str(request.get_port()), str(port))
+    return render(request, "host_port.html", {"new_host": host})
+
+
+def host_port_roundrobin_template(
+    request: HttpRequest, port1: int, port2: int, count: int = 1
+):
+    from reactpy_django import config
+
+    # Override ReactPy config to use round-robin hosts
+    original = config.REACTPY_DEFAULT_HOSTS
+    config.REACTPY_DEFAULT_HOSTS = cycle(
+        [
+            f"{request.get_host().split(':')[0]}:{port1}",
+            f"{request.get_host().split(':')[0]}:{port2}",
+        ]
+    )
+    html = render(
+        request,
+        "host_port_roundrobin.html",
+        {"count": range(max(count, 1))},
+    )
+
+    # Reset ReactPy config
+    config.REACTPY_DEFAULT_HOSTS = original
+    return html
 
 
 def view_to_component_sync_func(request):